Move css files right next to their components
