+++ src/App.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import './css/App.css';
+import './App.css';
 
 export class App extends Component {
   render() {
