+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import App from './App';
 
 import registerServiceWorker from './registerServiceWorker';
-import './css/index.css';
+import './index.css';
 
 ReactDOM.render(
     <App />,
